Refactor tests and add enum test case
