+++ compiler/ballerina-compiler-api/src/test/resources/test-src/symbol_at_cursor_basic_test.bal
