+++ compiler/ballerina-compiler-api/src/test/resources/test-src/symbol_at_cursor_enum_test.bal
@@ -1,0 +1,37 @@
+// Copyright (c) 2020 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+//
+// WSO2 Inc. licenses this file to you under the Apache License,
+// Version 2.0 (the "License"); you may not use this file except
+// in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+enum Colour {
+    RED, GREEN, BLUE
+}
+
+function testEnum() {
+    string red = RED;
+    Colour c = GREEN;
+    c = toColour("BLUE");
+}
+
+function toColour(string colour) returns Colour {
+    if (colour == RED) {
+        return RED;
+    } else if (colour == GREEN) {
+        return GREEN;
+    } else if (colour == BLUE) {
+        return BLUE;
+    }
+
+    panic error("Invalid colour: " + colour);
+}
